requirements.txt
pyTelegramBotAPI==4.15.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
